Add search bar and change component name to PickerMapComponent
