+++ frontend/package.json
@@ -10,14 +10,18 @@
     "dependencies": {
         "@vuepic/vue-datepicker": "^9.0.3",
         "axios": "^1.7.7",
+        "babel-polyfill": "^6.26.0",
         "core-js": "^3.8.3",
         "dompurify": "^3.1.7",
         "dotenv": "^16.4.5",
         "leaflet": "^1.9.4",
+        "leaflet-geosearch": "^4.0.0",
+        "leaflet-search": "^4.0.0",
         "marked": "^14.1.4",
         "marked-highlight": "^2.2.0",
         "path": "^0.12.7",
         "qrcode.vue": "^3.6.0",
+        "unfetch": "^5.0.0",
         "vue": "^3.5.11",
         "vue-router": "^4.4.5",
         "vue3-google-login": "^2.0.33",
