+++ frontend/src/views/ActivityCreate.vue
@@ -91,7 +91,7 @@
                 </div>
                 <label class="text-lg">Location</label>
                 <div class="flex justify-center">
-                    <MapComponent class="w-[75vw] h-[50vh]" />
+                    <MapComponent class="w-[75vw] h-[50vh] text-black" />
                 </div>
                 <div class="form-control w-full">
                     <div class="label">
@@ -212,7 +212,7 @@ import { isAuth, login } from '@/functions/Authentications';
 import { loadImage } from '@/functions/Utils';
 import ImageCarousel from '@/component/ImageCarousel';
 import { useRouter } from 'vue-router';
-import MapComponent from '@/component/MapComponent.vue';
+import MapComponent from '@/component/PickerMapComponent.vue';
 
 const MAX_IMAGE_COUNT = 10;
 const MAX_IMAGES_SIZE = 100e6; // 100 MB
